requirements.txt
Babel==2.15.0
contourpy==1.2.1
cycler==0.12.1
fonttools==4.52.4
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
psutil==5.9.8
pydub==0.25.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-tkdnd==0.2.1
scipy==1.13.1
six==1.16.0
tk==0.1.0
tkfilebrowser==2.3.2
ttkwidgets==0.13.0